requirements.txt
jinja2
google-cloud-storage